// frontend/src/api/generated/schemas/tradeOperationRead.ts
// @ts-nocheck
// This file is auto-generated by orval. Do not edit manually.
import type { TradeOperationReadEmotion } from "./tradeOperationReadEmotion";
import type { TradeOperationReadNotes } from "./tradeOperationReadNotes";
import type { TradeOperationReadOpTime } from "./tradeOperationReadOpTime";
import type { TradeOperationReadPrice } from "./tradeOperationReadPrice";
import type { TradeOperationReadQuantity } from "./tradeOperationReadQuantity";

export interface TradeOperationRead {
  created_at: string;
  emotion: TradeOperationReadEmotion;
  entry_id: string;
  id: string;
  notes: TradeOperationReadNotes;
  op_time: TradeOperationReadOpTime;
  op_type: string;
  price: TradeOperationReadPrice;
  quantity: TradeOperationReadQuantity;
  updated_at: string;
}